Domains can now be marked to be unreliable for mail deliverbility
If a domain is marked as such, users will be offered a PDF with the
lookup link in addition to the mail.

Why this feature?

Some mail providers (looking at you Microsoft!) have automatic systems
which can decide to put you on filter lists at any time for any
arbitrary reason.
(Isnt machine learning great?)

Anyway, using this feature, administrators can try to mitigate some of
the damage in case of deliverability problems.